slip out the executor function from the rule type definition
